Harden transfer domain and make the backend deployable to Render + Aiven, more details in description
Introduce production-grade reliability and security controls for the
transfer flow, then align the build and runtime configuration with a
containerized PaaS deployment.

Domain and persistence:
- Represent monetary values with BigDecimal (scale 2, HALF_EVEN) across the
  domain model, ports, events and use cases to eliminate double-precision
  rounding errors.
- Implement a double-entry ledger (immutable DEBIT/CREDIT postings) written in
  the same transaction as the balance update.
- Add a transactional outbox and a durable saga (instance/step persistence)
  so money movement and cross-service events survive crashes and restarts.
- Enforce API idempotency via a required Idempotency-Key, a Redis lock and a
  Postgres record, preventing duplicate transfers on client retries.

Security:
- Require authenticated requests on transfer endpoints, restrict admin routes to
  ROLE_ADMIN, and populate the SecurityContext from the client token; enforce
  account ownership on initiation. Map auth failures to 401/403 and conflicts
  to 409. Expose uid/roles in the auth endpoint.

Operations and deployment:
- Add a Flyway baseline, actuator health groups and graceful shutdown.
- Ship a multi-stage, non-root Dockerfile with dumb-init and a JVM healthcheck;
  make the build context-relative so it builds on Render.
- Add a Kafka-optional mode: when no broker is configured the outbox is
  processed locally, and the SASL mechanism is env-driven for Aiven.
- Add a Helm chart and base manifests (Postgres, Redis, Strimzi Kafka) for a
  future Kubernetes deployment.

+++ server/backend/src/main/java/com/intbank/core/domain/entity/Transfer.java
@@ -2,6 +2,7 @@ package com.intbank.core.domain.entity;
 
 import com.intbank.core.domain.vo.TransferStatus;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Objects;
 
@@ -11,7 +12,7 @@ public class Transfer
     private final String id;
     private final String fromAccountId;
     private final String toAccountId;
-    private final double amount;
+    private final BigDecimal amount;
     private final String currency;
     private final String reason;
     private TransferStatus status;
@@ -23,7 +24,7 @@ public class Transfer
         String id,
         String fromAccountId,
         String toAccountId,
-        double amount,
+        BigDecimal amount,
         String currency,
         String reason,
         Instant initiatedAt
@@ -63,7 +64,7 @@ public class Transfer
     public String id() { return id; }
     public String fromAccountId() { return fromAccountId; }
     public String toAccountId() { return toAccountId; }
-    public double amount() { return amount; }
+    public BigDecimal amount() { return amount; }
     public String currency() { return currency; }
     public String reason() { return reason; }
     public TransferStatus status() { return status; }
